moduleを切り出してconcernsの中に置く
