+++ sample_app/app/models/user.rb
@@ -24,20 +24,22 @@ class User < ApplicationRecord
 
 
   class << self
-    private
+    include UserConcern
 
-    def cost
-      ActiveModel::SecurePassword.min_cost ? BCrypt::Engine::MIN_COST : BCrypt::Engine.cost
-    end
-
-    public
-
-    def convert_string_to_digest(original_string)
-      BCrypt::Password.create(original_string, cost: cost)
-    end
-
-    def new_token
-      SecureRandom.urlsafe_base64
-    end
+    # private
+    #
+    # def cost
+    #   ActiveModel::SecurePassword.min_cost ? BCrypt::Engine::MIN_COST : BCrypt::Engine.cost
+    # end
+    #
+    # public
+    #
+    # def convert_string_to_digest(original_string)
+    #   BCrypt::Password.create(original_string, cost: cost)
+    # end
+    #
+    # def new_token
+    #   SecureRandom.urlsafe_base64
+    # end
   end
 end
